whisper_vad.py: input_features を model.dtype にキャスト
モデルが float16 でロードされている場合に float32 との dtype 不一致で
RuntimeError が発生するのを修正。

+++ scripts/whisper_vad.py
@@ -115,7 +115,7 @@ def transcribe_chunks(chunks, processor, model, device):
             task="transcribe",
             initial_prompt=INITIAL_PROMPT,
         )
-        input_features = inputs.input_features.to(device)
+        input_features = inputs.input_features.to(device, dtype=model.dtype)
 
         with torch.no_grad():
             generated_ids = model.generate(
